+++ docs/src/app/pages/components/Badge/ExampleBadges.js
@@ -11,7 +11,6 @@ const ExampleBadges = () => (
     <Badge text='6' theme='navy' />
     <Badge text='7' theme='border' />
     <Badge icon='icon-link-1' text='Word' theme='success' />
-    <Badge icon='icon-bell-1' theme='fog' />
   </div>
 )