+++ docs/src/app/pages/components/Checkbox/ExampleCheckboxChecked.js
@@ -2,7 +2,7 @@ import React from 'react'
 import Checkbox from 'react-conventions/lib/Checkbox'
 
 const ExampleCheckboxChecked = () => (
-  <Checkbox label="Checked checkbox" value={true} />
+  <Checkbox label='Checked checkbox' value={true} />
 )
 
 export default ExampleCheckboxChecked