+++ docs/src/app/pages/components/Checkbox/ExampleCheckboxLeft.js
@@ -2,7 +2,7 @@ import React from 'react'
 import Checkbox from 'react-conventions/lib/Checkbox'
 
 const ExampleCheckboxLeft = () => (
-  <Checkbox label="Checkbox with the label on the left side" value={true} labelPosition="left" />
+  <Checkbox label='Checkbox with the label on the left side' value={true} labelPosition='left' />
 )
 
 export default ExampleCheckboxLeft