+++ docs/src/app/pages/components/FileUpload/ExampleFileUploadCustomPreview.js
@@ -1,0 +1,8 @@
+import React from 'react'
+import FileUpload from 'react-conventions/lib/FileUpload'
+
+const ExampleFileUploadCustomPreview = () => (
+  <FileUpload label="File upload with custom preview size" showPreview={true} previewSize={100} />
+)
+
+export default ExampleFileUploadCustomPreview