+++ docs/src/app/pages/components/Toggle/ExampleToggleDefault.js
@@ -1,0 +1,8 @@
+import React from 'react'
+import Toggle from 'react-conventions/lib/Toggle'
+
+const ExampleToggleDefault = () => (
+  <Toggle label='Would you like to set this?' />
+)
+
+export default ExampleToggleDefault