+++ docs/src/app/pages/components/Toggle/ExampleToggleDisabled.js
@@ -1,0 +1,8 @@
+import React from 'react'
+import Toggle from 'react-conventions/lib/Toggle'
+
+const ExampleToggleDisabled = () => (
+  <Toggle label='Would you like to set this?' disabled={true} />
+)
+
+export default ExampleToggleDisabled