+++ docs/src/app/pages/components/Toggle/ExampleToggleOn.js
@@ -1,0 +1,8 @@
+import React from 'react'
+import Toggle from 'react-conventions/lib/Toggle'
+
+const ExampleToggleOn = () => (
+  <Toggle label='Would you like to set this?' value={true} />
+)
+
+export default ExampleToggleOn