+++ docs/src/app/pages/components/Toggle/ExampleToggleOptClass.js
@@ -1,0 +1,8 @@
+import React from 'react'
+import Toggle from 'react-conventions/lib/Toggle'
+
+const ExampleToggleOptClass = () => (
+  <Toggle label='Would you like to set this?' value={true} optClass='toggle-red' />
+)
+
+export default ExampleToggleOptClass