+++ docs/src/app/routes.js
@@ -14,6 +14,7 @@ import ButtonGroupPage from './pages/components/ButtonGroup/Page'
 import ColorsPage from './pages/foundations/Colors/Page'
 import ColorPickerPage from './pages/components/ColorPicker/Page'
 import CheckboxPage from './pages/components/Checkbox/Page'
+import DropdownPage from './pages/components/Dropdown/Page'
 import HomePage from './pages/Home/Page'
 import IconographyPage from './pages/foundations/Iconography/Page'
 import IconsPage from './pages/components/Icons/Page'
@@ -30,6 +31,8 @@ import ProgressBarPage from './pages/components/ProgressBar/Page'
 import PanelGroupPage from './pages/components/PanelGroup/Page'
 import SelectFieldPage from './pages/components/SelectField/Page'
 import TabsPage from './pages/components/TabWrapper/Page'
+import FileUploadPage from './pages/components/FileUpload/Page'
+import TogglePage from './pages/components/Toggle/Page'
 
 const Routes = (
   <Route path='/' component={Base}>
@@ -47,6 +50,7 @@ const Routes = (
       <Route path='alerts' title='Alerts' component={AlertsPage} />
       <Route path='badge' title='Badge' component={BadgePage} />
       <Route path="breadcrumb" title="Breadcrumb" component={BreadcrumbPage} />
+      <Route path='dropdown' title='Dropdown' component={DropdownPage} />
       <Route path='button-group' title='Button Group' component={ButtonGroupPage} />
       <Route path='buttons' title='Buttons' component={ButtonsPage} />
       <Route path='checkbox' title='Checkbox' component={CheckboxPage} />
@@ -65,6 +69,8 @@ const Routes = (
       <Route path='textarea' title='Textarea' component={TextareaPage} />
       <Route path='select-field' title='Select Field' component={SelectFieldPage} />
       <Route path='tooltip' title='Tooltip' component={TooltipPage} />
+      <Route path='file-upload' title='File Upload' component={FileUploadPage} />
+      <Route path='toggle' title='Toggle' component={TogglePage} />
     </Route>
     <Route path='patterns' title='Patterns'>
     </Route>