+++ CHANGELOG.md
@@ -1,5 +1,34 @@
 ## HEAD
 
+## 0.0.15
+###### _June 1, 2016_
+
+##### Feature
+- [Toggle] - ([#102](https://github.com/GetAmbassador/react-conventions/pull/102))
+
+##### Component Fixes / Enhancements
+- [FormGroup] - Prevent default on form submit
+([#106](https://github.com/GetAmbassador/react-conventions/pull/106))
+
+## 0.0.14
+###### _June 1, 2016_
+
+##### Feature
+- [Dropdown] - ([#103](https://github.com/GetAmbassador/react-conventions/pull/103))
+
+## 0.0.13
+###### _May 31, 2016_
+
+##### Component Fixes / Enhancements
+- [Badge] - Style improvements ([#101](https://github.com/GetAmbassador/react-conventions/pull/101))
+
+## 0.0.12
+###### _May 27, 2016_
+
+##### Component Fixes / Enhancements
+- [Badge] - Style improvements ([#99](https://github.com/GetAmbassador/react-conventions/pull/99))
+- [ActivityFeed] - Style improvements ([#97](https://github.com/GetAmbassador/react-conventions/pull/97))
+
 ## 0.0.11
 ###### _May 26, 2016_